requirements.txt
flask
langchain
langchain-core
langchain-community
langchain-chroma
unstructured
pypdf
lxml
beautifulsoup4
python-magic
filetype
pyyaml
diffusers
transformers
torch
accelerate
safetensors
sentence-transformers
Pillow
llama-cpp-python
markdown
ddgs